Add MultiCouplingFourSolvers1 and 2 according to new structure.
