+++ tests/tests.cmake
@@ -3,4 +3,8 @@
 #
 target_sources(testprecice
     PRIVATE
+    tests/serial/MultiCouplingFourSolvers1.cpp
+    tests/serial/MultiCouplingFourSolvers2.cpp
+    tests/serial/helpers.cpp
+    tests/serial/helpers.hpp
     )
